+++ PacketDotNet/Utils/Crc32.cs
@@ -15,15 +15,22 @@ using System.Text;
 namespace PacketDotNet.Utils
 {
     /// <summary>Implements a 32-bits cyclic redundancy check (CRC) hash algorithm.</summary>
-    /// <remarks>This class is not intended to be used for security purposes. For security applications use MD5, SHA1, SHA256, SHA384, 
-    /// or SHA512 in the System.Security.Cryptography namespace.</remarks>
+    /// <remarks>
+    ///     This class is not intended to be used for security purposes. For security applications use MD5, SHA1, SHA256,
+    ///     SHA384,
+    ///     or SHA512 in the System.Security.Cryptography namespace.
+    /// </remarks>
     public class Crc32 : HashAlgorithm
     {
         #region Fields
 
         /// <summary>Gets the default polynomial (used in WinZip, Ethernet, etc.)</summary>
-        /// <remarks>The default polynomial is a bit-reflected version of the standard polynomial 0x04C11DB7 used by WinZip, Ethernet, etc.</remarks>
+        /// <remarks>
+        ///     The default polynomial is a bit-reflected version of the standard polynomial 0x04C11DB7 used by WinZip,
+        ///     Ethernet, etc.
+        /// </remarks>
         public static readonly UInt32 DefaultPolynomial = 0xEDB88320; // Bitwise reflection of 0x04C11DB7;
+
         private const UInt32 AllOnes = 0xffffffff;
         private UInt32 _crc;
         private readonly UInt32[] _crc32Table;
@@ -34,22 +41,24 @@ namespace PacketDotNet.Utils
 
         #region Constructors
 
-        /// <summary>Creates a CRC32 object using the <see cref="DefaultPolynomial"/>.</summary>
+        /// <summary>Creates a CRC32 object using the <see cref="DefaultPolynomial" />.</summary>
         public Crc32()
-            : this((UInt32) DefaultPolynomial)
+            : this(DefaultPolynomial)
         {
         }
 
         /// <summary>Creates a CRC32 object using the specified polynomial.</summary>
-        /// <remarks>The polynomial should be supplied in its bit-reflected form. <see cref="DefaultPolynomial"/>.</remarks>
+        /// <remarks>The polynomial should be supplied in its bit-reflected form. <see cref="DefaultPolynomial" />.</remarks>
         public Crc32(UInt32 polynomial)
         {
             this.HashSizeValue = 32;
-            this._crc32Table = (UInt32[])CRC32TablesCache[polynomial];
-            if (this._crc32Table == null) {
+            this._crc32Table = (UInt32[]) CRC32TablesCache[polynomial];
+            if (this._crc32Table == null)
+            {
                 this._crc32Table = BuildCrc32Table(polynomial);
                 CRC32TablesCache.Add(polynomial, this._crc32Table);
             }
+
             this.Initialize();
         }
 
@@ -64,28 +73,28 @@ namespace PacketDotNet.Utils
 
         #region Public Methods
 
-        /// <summary>Computes the CRC32 value for the given ASCII string using the <see cref="DefaultPolynomial"/>.</summary>
+        /// <summary>Computes the CRC32 value for the given ASCII string using the <see cref="DefaultPolynomial" />.</summary>
         public static Int32 Compute(String asciiString)
         {
             DefaultCRC.Initialize();
             return ToInt32(DefaultCRC.ComputeHash(asciiString));
         }
 
-        /// <summary>Computes the CRC32 value for the given input stream using the <see cref="DefaultPolynomial"/>.</summary>
+        /// <summary>Computes the CRC32 value for the given input stream using the <see cref="DefaultPolynomial" />.</summary>
         public static Int32 Compute(Stream inputStream)
         {
             DefaultCRC.Initialize();
             return ToInt32(DefaultCRC.ComputeHash(inputStream));
         }
 
-        /// <summary>Computes the CRC32 value for the input data using the <see cref="DefaultPolynomial"/>.</summary>
+        /// <summary>Computes the CRC32 value for the input data using the <see cref="DefaultPolynomial" />.</summary>
         public static Int32 Compute(Byte[] buffer)
         {
             DefaultCRC.Initialize();
             return ToInt32(DefaultCRC.ComputeHash(buffer));
         }
 
-        /// <summary>Computes the hash value for the input data using the <see cref="DefaultPolynomial"/>.</summary>
+        /// <summary>Computes the hash value for the input data using the <see cref="DefaultPolynomial" />.</summary>
         public static Int32 Compute(Byte[] buffer, Int32 offset, Int32 count)
         {
             DefaultCRC.Initialize();
@@ -93,7 +102,10 @@ namespace PacketDotNet.Utils
         }
 
         /// <summary>Computes the hash value for the given ASCII string.</summary>
-        /// <remarks>The computation preserves the internal state between the calls, so it can be used for computation of a stream data.</remarks>
+        /// <remarks>
+        ///     The computation preserves the internal state between the calls, so it can be used for computation of a stream
+        ///     data.
+        /// </remarks>
         public Byte[] ComputeHash(String asciiString)
         {
             Byte[] rawBytes = Encoding.ASCII.GetBytes(asciiString);
@@ -101,26 +113,37 @@ namespace PacketDotNet.Utils
         }
 
         /// <summary>Computes the hash value for the given input stream.</summary>
-        /// <remarks>The computation preserves the internal state between the calls, so it can be used for computation of a stream data.</remarks>
+        /// <remarks>
+        ///     The computation preserves the internal state between the calls, so it can be used for computation of a stream
+        ///     data.
+        /// </remarks>
         public new Byte[] ComputeHash(Stream inputStream)
         {
             var buffer = new Byte[4096];
             Int32 bytesRead;
-            while ((bytesRead = inputStream.Read(buffer, 0, 4096)) > 0) {
+            while ((bytesRead = inputStream.Read(buffer, 0, 4096)) > 0)
+            {
                 this.HashCore(buffer, 0, bytesRead);
             }
+
             return this.HashFinal();
         }
 
         /// <summary>Computes the hash value for the input data.</summary>
-        /// <remarks>The computation preserves the internal state between the calls, so it can be used for computation of a stream data.</remarks>
+        /// <remarks>
+        ///     The computation preserves the internal state between the calls, so it can be used for computation of a stream
+        ///     data.
+        /// </remarks>
         public new Byte[] ComputeHash(Byte[] buffer)
         {
             return this.ComputeHash(buffer, 0, buffer.Length);
         }
 
         /// <summary>Computes the hash value for the input data.</summary>
-        /// <remarks>The computation preserves the internal state between the calls, so it can be used for computation of a stream data.</remarks>
+        /// <remarks>
+        ///     The computation preserves the internal state between the calls, so it can be used for computation of a stream
+        ///     data.
+        /// </remarks>
         public new Byte[] ComputeHash(Byte[] buffer, Int32 offset, Int32 count)
         {
             this.HashCore(buffer, offset, count);
@@ -140,7 +163,8 @@ namespace PacketDotNet.Utils
         /// <summary>Routes data written to the object into the hash algorithm for computing the hash.</summary>
         protected override void HashCore(Byte[] buffer, Int32 offset, Int32 count)
         {
-            for (Int32 i = offset; i < count; i++) {
+            for (Int32 i = offset; i < count; i++)
+            {
                 UInt64 ptr = (this._crc & 0xFF) ^ buffer[i];
                 this._crc >>= 8;
                 this._crc ^= this._crc32Table[ptr];
@@ -153,10 +177,10 @@ namespace PacketDotNet.Utils
             var finalHash = new Byte[4];
             UInt64 finalCrc = this._crc ^ AllOnes;
 
-            finalHash[3] = (Byte)((finalCrc >> 0) & 0xFF);
-            finalHash[2] = (Byte)((finalCrc >> 8) & 0xFF);
-            finalHash[1] = (Byte)((finalCrc >> 16) & 0xFF);
-            finalHash[0] = (Byte)((finalCrc >> 24) & 0xFF);
+            finalHash[3] = (Byte) ((finalCrc >> 0) & 0xFF);
+            finalHash[2] = (Byte) ((finalCrc >> 8) & 0xFF);
+            finalHash[1] = (Byte) ((finalCrc >> 16) & 0xFF);
+            finalHash[0] = (Byte) ((finalCrc >> 24) & 0xFF);
 
             return finalHash;
         }
@@ -171,14 +195,17 @@ namespace PacketDotNet.Utils
             var table = new UInt32[256];
 
             // 256 values representing ASCII character codes.
-            for (Int32 i = 0; i < 256; i++) {
-                var crc = (UInt32)i;
-                for (Int32 j = 8; j > 0; j--) {
+            for (Int32 i = 0; i < 256; i++)
+            {
+                var crc = (UInt32) i;
+                for (Int32 j = 8; j > 0; j--)
+                {
                     if ((crc & 1) == 1)
                         crc = (crc >> 1) ^ polynomial;
                     else
                         crc >>= 1;
                 }
+
                 table[i] = crc;
             }
